feat: :sparkles: rename files to .ts
