+++ src/module-of-darkness.ts
